Add footer
Add footer to bottom of page with links to Google Ads, Google Business
Solutions, Google Search, Google Privacy Policy, Google Terms of
Service, and Google Search Settings.

When attempting to change the background color of the footer, the color
didn't extend to the entire width of the viewport. This was because the
body had pre-defined margins due to the browser's default stylesheet.

To get away from the default stylesheet, Eric Meyer's "Reset CSS" 2.0
stylesheet is loaded before the main stylesheet (renamed to main.css).
The CSS reset removed the body margins and footer's background color
stretched the width of the viewport as expected.

Loading the CSS reset (reset.css) affected the perceived margins and
paddings of other elements since they didn't take into account the body
margins from the default stylesheet. Thus those margins and paddings
were set again to ensure the spacing looked the same as before.